prep for mocking ocr - moving stuff out of api.py
