+++ app/api.py
@@ -1,19 +1,14 @@
-import io
-import re
-import warnings
 from datetime import datetime
-from typing import List
 
-import easyocr
 from fastapi import APIRouter, Depends, File, UploadFile, status
 from fastapi.responses import JSONResponse
-from PIL import Image
 from pydantic import BaseModel, ConfigDict
-from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from app.database import get_db
-from app.models import Token, Transaction
+from app.logic.transactions import process_add_transaction
+from app.models import Token
+from app.ocr import extract_transactions_from_image_upload
 
 router = APIRouter(prefix="/api", tags=["API"])
 
@@ -64,106 +59,6 @@ class TransactionCreate(BaseModel):
             }
         }
     )
-
-
-def process_add_transaction(
-    timestamp: datetime,
-    from_token: str,
-    to_token: str,
-    from_amount: float,
-    to_amount: float,
-    db: Session,
-):
-    """
-    Process and validate a transaction between two tokens.
-
-    This function implements the business logic for token transactions:
-    - Validates that both tokens exist in the database
-    - Ensures that exactly one token is a stablecoin
-    - Calculates final USD values and token amounts
-    - Stores the transaction in the database
-
-    Args:
-        timestamp (datetime): When the transaction occurred
-        from_token (str): The source token symbol
-        to_token (str): The destination token symbol
-        from_amount (float): Amount of the source token
-        to_amount (float): Amount of the destination token
-        db (Session): Database session for querying and saving
-
-    Returns:
-        dict or JSONResponse: Success message on success, or error response on failure
-
-    Raises:
-        IntegrityError: If a transaction with the same token and timestamp already exists
-    """
-    # Validate that tokens exist and get their stability status
-    from_token_obj = db.query(Token).filter(Token.name == from_token).first()
-    to_token_obj = db.query(Token).filter(Token.name == to_token).first()
-
-    if not from_token_obj:
-        return JSONResponse(
-            content={
-                "error": f"'{from_token}' is not recognized. Please add it first."
-            },
-            status_code=status.HTTP_400_BAD_REQUEST,
-        )
-    if not to_token_obj:
-        return JSONResponse(
-            content={"error": f"'{to_token}' is not recognized. Please add it first."},
-            status_code=status.HTTP_400_BAD_REQUEST,
-        )
-    if from_token_obj.is_stable and to_token_obj.is_stable:
-        return JSONResponse(
-            content={"error": "Both tokens cannot be stablecoins"},
-            status_code=status.HTTP_400_BAD_REQUEST,
-        )
-    if not from_token_obj.is_stable and not to_token_obj.is_stable:
-        return JSONResponse(
-            content={"error": "One of the tokens must be a stablecoin"},
-            status_code=status.HTTP_400_BAD_REQUEST,
-        )
-
-    # Determine which is the stablecoin and which is the non-stablecoin
-    if from_token_obj.is_stable:
-        stablecoin = from_token
-        non_stablecoin = to_token
-        final_usd = -from_amount
-        final_amount = to_amount
-    else:
-        stablecoin = to_token
-        non_stablecoin = from_token
-        final_usd = to_amount
-        final_amount = -from_amount
-
-    try:
-        new_transaction = Transaction(
-            timestamp=timestamp,
-            token=non_stablecoin,
-            amount=final_amount,
-            stable_coin=stablecoin,
-            total_usd=final_usd,
-        )
-        db.add(new_transaction)
-        db.commit()
-    except IntegrityError:
-        db.rollback()
-        return JSONResponse(
-            content={
-                "error": f"Transaction for '{non_stablecoin}' at '{timestamp}' already exists."
-            },
-            status_code=status.HTTP_409_CONFLICT,
-        )
-
-    return {
-        "status": "success",
-        "timestamp": timestamp,
-        "token": non_stablecoin,
-        "amount": final_amount,
-        "stable_coin": stablecoin,
-        "total_usd": final_usd,
-        "message": f"Transaction added: timestamp '{timestamp}', token '{non_stablecoin}', amount '{final_amount}', stable_coin '{stablecoin}', total_usd '{final_usd}'.",
-    }
 
 
 @router.post(
@@ -352,14 +247,6 @@ async def get_token(token_name: str, db: Session = Depends(get_db)):
     return {"name": token.name, "is_stable": token.is_stable}
 
 
-class ExtractedTransaction(BaseModel):
-    timestamp: datetime
-    from_token: str
-    to_token: str
-    from_amount: float
-    to_amount: float
-
-
 @router.post("/transactions/extract", response_class=JSONResponse)
 async def extract_transactions_from_image(
     image: UploadFile = File(...),
@@ -376,135 +263,11 @@ async def extract_transactions_from_image(
         )
 
     try:
-        # Read the image
-        contents = await image.read()
-        img = Image.open(io.BytesIO(contents))
 
-        # Suppress the specific pin_memory warning
-        warnings.filterwarnings("ignore", message=".*pin_memory.*no accelerator.*")
-
-        # Extract text using OCR
-        reader = easyocr.Reader(["en"], download_enabled=False, gpu=False)
-        result = reader.readtext(img)
-        extracted_text = " ".join([text[1] for text in result])
-
-        # Parse the text to extract transaction data
-        transactions = parse_debank_screenshot(extracted_text)
-
-        if not transactions:
-            return JSONResponse(
-                content={
-                    "status": "info",
-                    "message": "No transactions found in the image. Extracted: "
-                    + extracted_text,
-                },
-                status_code=200,
-            )
-
-        # Process and save each transaction
-        results = []
-        for t in transactions:
-            result = process_add_transaction(
-                timestamp=t.timestamp,
-                from_token=t.from_token,
-                to_token=t.to_token,
-                from_amount=t.from_amount,
-                to_amount=t.to_amount,
-                db=db,
-            )
-            results.append(result)
-
-        # Count successful transactions
-        successful = sum(
-            1 for r in results if isinstance(r, dict) and r.get("status") == "success"
-        )
-
-        return {
-            "status": "success" if successful > 0 else "info",
-            "message": f"Added {successful} out of {len(transactions)} transactions from the image.",
-            "details": results,
-        }
+        return await extract_transactions_from_image_upload(image, db)
 
     except Exception as e:
         return JSONResponse(
             content={"error": f"Failed to process image: {str(e)}"},
             status_code=500,
         )
-
-
-def parse_debank_screenshot(text: str) -> List[ExtractedTransaction]:
-    """
-    Parse text extracted from a Debank screenshot to identify transactions.
-    """
-
-    transactions = []
-    text = text.replace("\n", " ").replace("\r", " ")
-
-    # Split by "Contract Interaction"
-    sections = text.split("Contract Interaction")
-
-    for i, section in enumerate(sections[1:], 1):
-        section = section.strip()
-        if not section:
-            continue
-
-        curr_transaction = {}
-
-        # More flexible regex patterns to handle OCR errors
-        from_patterns = [
-            r"-\s*(\d+(?:\.\d+)?)\s+([A-Z]+)\s*\([s$]?[\d,.]+\)",
-            r"-(\d+(?:\.\d+)?)\s+([A-Z]+)",
-        ]
-
-        for pattern in from_patterns:
-            from_match = re.search(pattern, section)
-            if from_match:
-                curr_transaction["from_amount"] = float(from_match.group(1))
-                curr_transaction["from_token"] = from_match.group(2)
-                break
-
-        to_patterns = [
-            r"\+\s*(\d+(?:\.\d+)?)\s+([A-Z]+)\s*\(\$[\d,.]+\)",
-            r"\+(\d+(?:\.\d+)?)\s+([A-Z]+)",
-        ]
-
-        for pattern in to_patterns:
-            to_match = re.search(pattern, section)
-            if to_match:
-                curr_transaction["to_amount"] = float(to_match.group(1))
-                curr_transaction["to_token"] = to_match.group(2)
-                break
-
-        # Handle timestamp with flexible separators
-        timestamp_patterns = [
-            r"(\d{4}/\d{2}/\d{2})\s+(\d{2})[.:](\d{2})[.:](\d{2})",
-            r"(\d{4}/\d{2}/\d{2})\s+(\d{1,2})[.:](\d{2})[.:](\d{2})",
-        ]
-
-        for pattern in timestamp_patterns:
-            timestamp_match = re.search(pattern, section)
-            if timestamp_match:
-                try:
-                    date_part = timestamp_match.group(1)
-                    hour = timestamp_match.group(2).zfill(2)
-                    minute = timestamp_match.group(3)
-                    second = timestamp_match.group(4)
-                    timestamp_str = f"{date_part} {hour}:{minute}:{second}"
-                    curr_transaction["timestamp"] = datetime.strptime(
-                        timestamp_str, "%Y/%m/%d %H:%M:%S"
-                    )
-                    break
-                except ValueError:
-                    continue
-
-        # If we have all required fields, add the transaction
-        if all(
-            k in curr_transaction
-            for k in ["timestamp", "from_token", "to_token", "from_amount", "to_amount"]
-        ):
-            try:
-                transactions.append(ExtractedTransaction(**curr_transaction))
-            except Exception:
-                pass
-
-    return transactions
